addeed ci and tacm playwright tests, restructure playwright utils
