+++ tests/playwright/reports/ci.spec.ts
@@ -1,0 +1,28 @@
+import { test } from "@playwright/test";
+import { stateUserAuthPath } from "../utils/consts";
+import {
+  navigateToReportHome,
+  navigateToAddEditReportModal,
+  fillAddEditReportModal,
+  assertReportIsCreated,
+  testModalData,
+} from "./../utils/reportUtils";
+
+test.use({ storageState: stateUserAuthPath });
+
+const reportSpecificData = {
+  reportButtonName: "Enter CI Report online",
+  startReportButtonName: "Start Critical Incident Report",
+  modalHeading: "Add new Critical Incident Report",
+  reportNameInputHeading: "Critical Incident Report Name",
+};
+
+test("create a CI report as a state user", async ({ page }) => {
+  await navigateToReportHome(page, reportSpecificData.reportButtonName);
+  await navigateToAddEditReportModal(
+    page,
+    reportSpecificData.startReportButtonName
+  );
+  await fillAddEditReportModal(page, reportSpecificData);
+  await assertReportIsCreated(page, testModalData);
+});
